% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_population_density_column}
\alias{add_population_density_column}
\title{Add Population Density Column to Country Data}
\usage{
add_population_density_column(
  df,
  id_column,
  id_type = "iso3_code",
  date_column = NULL,
  target_column = "population_density"
)
}
\description{
Add Population Density Column to Country Data
}
