% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_population_share_column}
\alias{add_population_share_column}
\title{Add Population Share Column to Country Data}
\usage{
add_population_share_column(
  df,
  id_column,
  id_type = "regex",
  date_column = NULL,
  value_column = "value",
  target_column = "population_share",
  include_estimates = FALSE
)
}
\description{
Add Population Share Column to Country Data
}
