% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_poverty_ratio_column}
\alias{add_poverty_ratio_column}
\title{Add Poverty Ratio Column to Country Data}
\usage{
add_poverty_ratio_column(
  df,
  id_column,
  id_type = "iso3_code",
  date_column = NULL,
  target_column = "poverty_ratio"
)
}
\description{
Add Poverty Ratio Column to Country Data
}
