% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econtools-package.R
\docType{package}
\name{econtools-package}
\alias{econtools}
\alias{econtools-package}
\title{econtools: Enrich and Analyze Sovereign-Level Economic Data}
\description{
Provides a consistent set of functions for enriching and analyzing sovereign-level economic data. Economists, data scientists, and financial professionals can use the package to add standardized identifiers, demographic and macroeconomic indicators, and derived metrics such as gross domestic product per capita or government expenditure shares.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidy-intelligence/r-econtools}
  \item \url{https://tidy-intelligence.github.io/r-econtools/}
  \item Report bugs at \url{https://github.com/tidy-intelligence/r-econtools/issues}
}

}
\author{
\strong{Maintainer}: Christoph Scheuch \email{christoph@tidy-intelligence.com} (\href{https://orcid.org/0009-0004-0423-6819}{ORCID}) [copyright holder]

}
\keyword{internal}
