% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econtools-package.R
\docType{package}
\name{econtools-package}
\alias{econtools}
\alias{econtools-package}
\title{econtools: Tools to Analyze Economic Data}
\description{
A collection of tools to facilitate the analysis of economic data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidy-intelligence/r-econtools}
  \item \url{https://tidy-intelligence.github.io/r-econtools/}
  \item Report bugs at \url{https://github.com/tidy-intelligence/r-econtools/issues}
}

}
\author{
\strong{Maintainer}: Christoph Scheuch \email{christoph@tidy-intelligence.com} (\href{https://orcid.org/0009-0004-0423-6819}{ORCID}) [copyright holder]

}
\keyword{internal}
